+++ R/download_cloudmask.R
@@ -2,11 +2,12 @@
 #'
 #' @param aoi geometry corresponding to one point
 #' @param raster_dir directory where rasters are stored
-#' @param collection_info description of collection to download
+#' @param item_collection description of collection to download
 #' @param iChar plot ID
 #' @param collection character. name of the collection
 #' @param asset_names character.
 #' @param resolution numeric. spatial resolution (10 or 20)
+#' @param siteName character. name of the study site
 #'
 #' @return list of collections per plot
 #' @importFrom rstac assets_url
@@ -14,72 +15,65 @@
 #' @importFrom sf st_transform st_bbox
 #' @export
 #'
-download_cloudmask <- function(aoi, raster_dir, collection_info, iChar,
+download_cloudmask <- function(aoi, raster_dir, item_collection, iChar,
                                collection = 'sentinel-2-l2a', asset_names,
-                               resolution){
+                               resolution, siteName = NULL){
 
-  cloud_dl <- out_dir <- NULL
   # if collection not empty
-  if (length(collection_info$acquisitionDate)>0){
+  if (length(item_collection$acquisitionDate)>0){
     # stac_source <- rstac::stac("https://planetarycomputer.microsoft.com/api/stac/v1")
-    # get bounding box
-    plots_bbox <- aoi |>
-      # sf::st_transform(4326) |>
-      sf::st_bbox()
-    # get all SCL assets
-    band_url <- rstac::assets_url(items = collection_info, asset_names = asset_names)
-    dateAcq <- get_dateAcq(band_url = band_url, collection = collection)
-    if (collection == 'sentinel-2-l2a'){
-      stac_query <- collection_info |>
-        rstac::items_sign(
-          rstac::sign_planetary_computer())
-      band_url <- rstac::assets_url(items = stac_query,
-                                    asset_names = asset_names,
-                                    append_gdalvsi = T)
-      # band_url <- as.list(make_vsicurl_url(band_url, collection = collection))
-    } else if (collection == 'sentinel2-l2a-sen2lasrc'){
-      band_url <- as.list(make_vsicurl_theia_url(band_url))
-    }
     # define file name for cloud mask
-    filenames <- basename(paste0('plot_',iChar,'_',dateAcq,'_',asset_names,'.tiff'))
+    S2product <- unlist(lapply(item_collection$features,'[[','id'))
+    dateAcq <- get_dateAcq(S2product = S2product, collection = collection)
+    if (is.null(siteName)){
+      filenames <- basename(paste0('plot_',iChar,'_',dateAcq,'_',asset_names,'.tiff'))
+    } else {
+      filenames <- basename(paste0(siteName,'_',iChar,'_',dateAcq,'_',asset_names,'.tiff'))
+    }
     filepath <- file.path(raster_dir,filenames)
     # check if expected files already exist
     exist_file <- file.exists(filepath)
-    # define temporary directory for plot SCL
-    out_dir <- file.path(raster_dir,paste0('tmp_',asset_names,'_',iChar))
-    filepath_tmp <- file.path(out_dir,filenames)
-    filepath_consolidate <- filepath
     # if cloud file does not exists
+    features_dl <- features_exist <- NULL
     if (FALSE %in% exist_file){
-      dir.create(out_dir, showWarnings = F, recursive = T)
       # keep downloading these files
       keep_dl <- which(exist_file==FALSE)
       # corresponding acquisition date
       keep_dateAcq <- dateAcq[keep_dl]
-      out_file <-  as.list(file.path(out_dir,filenames[keep_dl]))
-      band_url <- band_url[keep_dl]
+      item_collection <- item_collection$features[keep_dl]
       # download
-      mapply(FUN = get_asset, asset_url = band_url, out_file = out_file,
-             MoreArgs = list(plots_bbox = plots_bbox))
-      filepath_consolidate[keep_dl] <- out_file
+      features_dl <- lapply(X = item_collection, 
+                            FUN = get_asset_terra, 
+                            asset_names = asset_names, 
+                            collection = collection, 
+                            aoi = aoi)
+      names(features_dl) <- keep_dateAcq
+      for (i in seq_len(length(features_dl)))
+        features_dl[[i]] <- features_dl[[i]][[1]]
     }
-    cloud_dl <- lapply(X = filepath_consolidate, FUN = terra::rast)
+    if (TRUE %in% exist_file){
+      read_exists <- which(exist_file==TRUE)
+      features_exist <- lapply(filepath, terra::rast)
+      names(features_exist) <- dateAcq[read_exists]
+      item_collection <- item_collection$features
+    }
+    if (is.null(features_exist)) cloudmask <- features_dl
+    if (is.null(features_dl)) cloudmask <- features_exist
+    if (! is.null(features_dl) & ! is.null(features_exist)) 
+      cloudmask <- c(features_exist, features_dl)
+
     # get B2 asset as template for 10m band if needed
     if (resolution == 10 & asset_names == 'SCL') {
-      asset_b2 <- 'B02'
-      stac_query <- collection_info |>
-        rstac::items_sign(
-          rstac::sign_planetary_computer())
-      b02_url <- rstac::assets_url(items = stac_query,
-                                   asset_names = asset_b2,
-                                   append_gdalvsi = T)
-      # b02_url <- make_vsicurl_url(b02_url, collection = collection)[1]
-      out_file <- tempfile()
-      get_asset(asset_url = b02_url, out_file = out_file, plots_bbox = plots_bbox)
-      temp_10m <- terra::rast(out_file)
-      cloud_dl  <- lapply(X = cloud_dl, FUN = terra::resample, temp_10m, method = 'near')
+      asset_b2 <- 'B02' 
+      features_b2 <- get_asset_terra(item = item_collection[[1]], 
+                                     asset_names = asset_b2, 
+                                     collection = collection, 
+                                     aoi = aoi)
+      cloudmask  <- lapply(X = cloudmask, 
+                          FUN = terra::resample, 
+                          y = features_b2[[1]], 
+                          method = 'near')
     }
-    names(cloud_dl) <- as.Date(dateAcq)
   }
-  return(list('cloud_dl' = cloud_dl, 'out_dir' = out_dir))
+  return(cloudmask)
 }